% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpra_5locus.R
\name{cpra_5locus}
\alias{cpra_5locus}
\title{FUNCTION_TITLE}
\usage{
cpra_5locus(excludes, DT, ethnic.weights)
}
\arguments{
\item{excludes}{PARAM_DESCRIPTION}

\item{DT}{PARAM_DESCRIPTION}

\item{ethnic.weights}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
